don't fail delete renumbered if files doesn't exist

+++ strucmotif-search-core/src/main/java/org/rcsb/strucmotif/io/StructureDataProviderImpl.java
@@ -209,9 +209,12 @@ public class StructureDataProviderImpl implements StructureDataProvider {
     }
 
     @Override
-    public void removeRenumbered(StructureIdentifier structureIdentifier) {
+    public void deleteRenumbered(StructureIdentifier structureIdentifier) {
         try {
-            Files.delete(getRenumberedStructurePath(structureIdentifier));
+            Path renumberedPath = getRenumberedStructurePath(structureIdentifier);
+            if (Files.exists(renumberedPath)) {
+                Files.delete(getRenumberedStructurePath(structureIdentifier));
+            }
         } catch (IOException e) {
             throw new UncheckedIOException(e);
         }
